requirements.txt
asgiref            3.8.1
Django             5.1.1
django-jalali      7.1.0
django-jalali-date 1.1.3
jalali_core        1.0.0
jdatetime          5.0.0
pip                22.0.2
setuptools         59.6.0
sqlparse           0.5.1
typing_extensions  4.12.2
